added lr loss tests
